Implemented the setup  view with the color themere in a scrollable view. TODO: Other seutp modules should be added.
