+++ Arduino Telnet Chess/GamesView.swift
@@ -9,9 +9,7 @@ import SwiftUI
 import CoreData
 
 struct GamesView: View {
-    
     @Environment(\.managedObjectContext) private var viewContext
-    
     @FetchRequest(
         sortDescriptors: [NSSortDescriptor(keyPath: \Item.timestamp, ascending: true)],
         animation: .default)
